extract common functionality in tests
